use decimal over double

+++ LeagueOptimizer.Abstractions/Champions/Data/AttackSpeedData.cs
@@ -2,5 +2,5 @@ namespace LeagueOptimizer.Abstractions.Champions.Data;
 
 public class AttackSpeedData : StatData
 {
-    public double Ratio { get; set; }
+    public decimal Ratio { get; set; }
 }
